Admin: refactor user and course detail pages into sections
Extract self-contained section components from monolithic detail pages,
rename ViewPage to Page, and inline simple table components.
