+++ judgels-client/src/components/forms/FormTableInput/FormTableInput.jsx
@@ -6,10 +6,19 @@ import { getIntent } from '../meta';
 
 import './FormTableInput.scss';
 
-export function FormTableInput({ className, keyClassName, label, labelHelper, meta, inputHelper, children }) {
+export function FormTableInput({
+  className,
+  keyStyles,
+  keyClassName,
+  label,
+  labelHelper,
+  meta,
+  inputHelper,
+  children,
+}) {
   return (
     <tr className="form-table-input">
-      <td className={keyClassName}>
+      <td style={keyStyles} className={keyClassName}>
         <div className="form-table-input__group-label">
           <span className="form-table-input__label">{label}</span>
           {labelHelper && <span className="form-table-input__label-helper"> ({labelHelper})</span>}
