--- judgels-client/src/routes/admin/courses/CourseChaptersTable/CourseChaptersTable.scss
@@ -1,5 +1,0 @@
-.course-chapters-table {
-  .col-alias {
-    width: 50px;
-  }
-}
