+++ judgels-client/src/routes/admin/courses/CourseGeneralSection/CourseGeneralSection.jsx
@@ -1,0 +1,72 @@
+import { Button, Intent } from '@blueprintjs/core';
+import { Edit } from '@blueprintjs/icons';
+import { Flex } from '@blueprintjs/labs';
+import { useMutation } from '@tanstack/react-query';
+import { useState } from 'react';
+
+import { FormTable } from '../../../../components/forms/FormTable/FormTable';
+import { updateCourseMutationOptions } from '../../../../modules/queries/course';
+import CourseGeneralEditForm from '../CourseGeneralEditForm/CourseGeneralEditForm';
+
+import * as toastActions from '../../../../modules/toast/toastActions';
+
+export function CourseGeneralSection({ course }) {
+  const updateCourseMutation = useMutation(updateCourseMutationOptions(course.jid));
+
+  const [isEditing, setIsEditingGeneral] = useState(false);
+
+  const keyStyles = { width: '250px' };
+
+  const detailRows = [
+    { key: 'slug', title: 'Slug', value: course.slug },
+    { key: 'name', title: 'Name', value: course.name },
+    { key: 'description', title: 'Description', value: course.description },
+  ];
+
+  const updateCourse = async data => {
+    await updateCourseMutation.mutateAsync(data, {
+      onSuccess: () => toastActions.showSuccessToast('Course updated.'),
+    });
+    setIsEditingGeneral(false);
+  };
+
+  const renderEditButton = () => {
+    return (
+      !isEditing && (
+        <Button small intent={Intent.PRIMARY} icon={<Edit />} onClick={() => setIsEditingGeneral(true)}>
+          Edit
+        </Button>
+      )
+    );
+  };
+
+  const renderContent = () => {
+    if (isEditing) {
+      const initialValues = {
+        slug: course.slug || '',
+        name: course.name || '',
+        description: course.description || '',
+      };
+      return (
+        <CourseGeneralEditForm
+          initialValues={initialValues}
+          onSubmit={updateCourse}
+          onCancel={() => setIsEditingGeneral(false)}
+        />
+      );
+    }
+    return <FormTable keyStyles={keyStyles} rows={detailRows} />;
+  };
+
+  return (
+    <div>
+      <Flex asChild justifyContent="space-between" alignItems="baseline">
+        <h4>
+          <span>General</span>
+          {renderEditButton()}
+        </h4>
+      </Flex>
+      {renderContent()}
+    </div>
+  );
+}
