--- judgels-client/src/routes/admin/courses/CoursesTable/CoursesTable.scss
@@ -1,9 +1,0 @@
-.courses-table {
-  .col-id {
-    width: 60px;
-  }
-
-  .col-slug {
-    width: 200px;
-  }
-}
