+++ judgels-client/src/routes/admin/users/UsersPage/UsersPage.jsx
@@ -1,13 +1,14 @@
+import { HTMLTable } from '@blueprintjs/core';
 import { useQuery } from '@tanstack/react-query';
-import { useLocation } from '@tanstack/react-router';
+import { Link, useLocation } from '@tanstack/react-router';
 
 import { ActionButtons } from '../../../../components/ActionButtons/ActionButtons';
 import { ContentCard } from '../../../../components/ContentCard/ContentCard';
+import { FormattedDate } from '../../../../components/FormattedDate/FormattedDate';
 import { LoadingState } from '../../../../components/LoadingState/LoadingState';
 import Pagination from '../../../../components/Pagination/Pagination';
 import { usersQueryOptions } from '../../../../modules/queries/user';
 import { UserUpsertDialog } from '../UserUpsertDialog/UserUpsertDialog';
-import { UsersTable } from '../UsersTable/UsersTable';
 
 const PAGE_SIZE = 250;
 
@@ -39,7 +40,28 @@ export default function UsersPage() {
       );
     }
 
-    return <UsersTable users={data.page} lastSessionTimesMap={lastSessionTimesMap} />;
+    const rows = data.page.map(user => (
+      <tr key={user.jid}>
+        <td>
+          <Link to={`/admin/users/${user.username}`}>{user.username}</Link>
+        </td>
+        <td>{user.email}</td>
+        <td>{lastSessionTimesMap[user.jid] ? <FormattedDate value={lastSessionTimesMap[user.jid]} /> : '-'}</td>
+      </tr>
+    ));
+
+    return (
+      <HTMLTable striped className="table-list">
+        <thead>
+          <tr>
+            <th>Username</th>
+            <th>Email</th>
+            <th>Last login</th>
+          </tr>
+        </thead>
+        <tbody>{rows}</tbody>
+      </HTMLTable>
+    );
   };
 
   return (
